+++ src/core/Store/DataInput.cs
@@ -154,33 +154,33 @@ namespace Lucene.Net.Store
             return i;
 
             /*byte b = ReadByte();
-            if (b >= 0)
+            if ((sbyte)b >= 0)
             {
                 return b;
             }
             int i = b & 0x7F;
             b = ReadByte();
             i |= (b & 0x7F) << 7;
-            if (b >= 0)
+            if ((sbyte)b >= 0)
             {
                 return i;
             }
             b = ReadByte();
             i |= (b & 0x7F) << 14;
-            if (b >= 0)
+            if ((sbyte)b >= 0)
             {
                 return i;
             }
             b = ReadByte();
             i |= (b & 0x7F) << 21;
-            if (b >= 0)
+            if ((sbyte)b >= 0)
             {
                 return i;
             }
             b = ReadByte();
             // Warning: the next ands use 0x0F / 0xF0 - beware copy/paste errors:
             i |= (b & 0x0F) << 28;
-            if ((b & 0xF0) == 0)
+            if (((sbyte)b & 0xF0) == 0)
             {
                 return i;
             }
